frontend/requirements.txt
chainlit
httpx
python-dotenv